Turn production predicates into default trait impl
