--- light-client/src/predicates/production.rs
@@ -1,213 +1,0 @@
-use crate::prelude::*;
-
-use tendermint::lite::ValidatorSet as _;
-
-#[derive(Copy, Clone, Debug)]
-pub struct ProdPredicates;
-
-impl VerificationPredicates for ProdPredicates {
-    fn validator_sets_match(&self, light_block: &LightBlock) -> Result<(), VerificationError> {
-        ensure!(
-            light_block.signed_header.header.validators_hash == light_block.validators.hash(),
-            VerificationError::InvalidValidatorSet {
-                header_validators_hash: light_block.signed_header.header.validators_hash,
-                validators_hash: light_block.validators.hash(),
-            }
-        );
-
-        Ok(())
-    }
-
-    fn next_validators_match(&self, light_block: &LightBlock) -> Result<(), VerificationError> {
-        ensure!(
-            light_block.signed_header.header.next_validators_hash
-                == light_block.next_validators.hash(),
-            VerificationError::InvalidNextValidatorSet {
-                header_next_validators_hash: light_block.signed_header.header.next_validators_hash,
-                next_validators_hash: light_block.next_validators.hash(),
-            }
-        );
-
-        Ok(())
-    }
-
-    fn header_matches_commit(
-        &self,
-        signed_header: &SignedHeader,
-        header_hasher: &dyn HeaderHasher,
-    ) -> Result<(), VerificationError> {
-        let header_hash = header_hasher.hash(&signed_header.header);
-
-        ensure!(
-            header_hash == signed_header.commit.block_id.hash,
-            VerificationError::InvalidCommitValue {
-                header_hash,
-                commit_hash: signed_header.commit.block_id.hash,
-            }
-        );
-
-        Ok(())
-    }
-
-    fn valid_commit(
-        &self,
-        commit: &Commit,
-        validators: &ValidatorSet,
-        validator: &dyn CommitValidator,
-    ) -> Result<(), VerificationError> {
-        // FIXME: Do not discard underlying error
-        validator
-            .validate(commit, validators)
-            .map_err(|e| VerificationError::InvalidCommit(e.to_string()))?;
-
-        Ok(())
-    }
-
-    fn is_within_trust_period(
-        &self,
-        header: &Header,
-        trusting_period: Duration,
-        now: Time,
-    ) -> Result<(), VerificationError> {
-        let expires_at = header.time + trusting_period;
-
-        ensure!(
-            header.time < now && expires_at > now,
-            VerificationError::NotWithinTrustPeriod {
-                at: expires_at,
-                now,
-            }
-        );
-
-        ensure!(
-            header.time <= now,
-            VerificationError::HeaderFromTheFuture {
-                header_time: header.time,
-                now
-            }
-        );
-
-        Ok(())
-    }
-
-    fn is_monotonic_bft_time(
-        &self,
-        untrusted_header: &Header,
-        trusted_header: &Header,
-    ) -> Result<(), VerificationError> {
-        ensure!(
-            untrusted_header.time > trusted_header.time,
-            VerificationError::NonMonotonicBftTime {
-                header_bft_time: untrusted_header.time,
-                trusted_header_bft_time: trusted_header.time,
-            }
-        );
-
-        Ok(())
-    }
-
-    fn is_monotonic_height(
-        &self,
-        untrusted_header: &Header,
-        trusted_header: &Header,
-    ) -> Result<(), VerificationError> {
-        let trusted_height: Height = trusted_header.height.into();
-
-        ensure!(
-            untrusted_header.height > trusted_header.height,
-            VerificationError::NonIncreasingHeight {
-                got: untrusted_header.height.into(),
-                expected: trusted_height + 1,
-            }
-        );
-
-        Ok(())
-    }
-
-    fn has_sufficient_voting_power(
-        &self,
-        signed_header: &SignedHeader,
-        validators: &ValidatorSet,
-        trust_threshold: &TrustThreshold,
-        calculator: &dyn VotingPowerCalculator,
-    ) -> Result<(), VerificationError> {
-        // FIXME: Do not discard underlying error
-        let total_power = calculator.total_power_of(validators);
-        let voting_power = calculator
-            .voting_power_in(signed_header, validators)
-            .map_err(|e| VerificationError::ImplementationSpecific(e.to_string()))?;
-
-        ensure!(
-            voting_power * trust_threshold.denominator > total_power * trust_threshold.numerator,
-            VerificationError::InsufficientVotingPower {
-                total_power,
-                voting_power,
-            }
-        );
-
-        Ok(())
-    }
-
-    fn has_sufficient_validators_overlap(
-        &self,
-        untrusted_sh: &SignedHeader,
-        trusted_validators: &ValidatorSet,
-        trust_threshold: &TrustThreshold,
-        calculator: &dyn VotingPowerCalculator,
-    ) -> Result<(), VerificationError> {
-        // FIXME: Do not discard underlying error
-        let total_power = calculator.total_power_of(trusted_validators);
-        let voting_power = calculator
-            .voting_power_in(untrusted_sh, trusted_validators)
-            .map_err(|e| VerificationError::ImplementationSpecific(e.to_string()))?;
-
-        ensure!(
-            voting_power * trust_threshold.denominator > total_power * trust_threshold.numerator,
-            VerificationError::InsufficientValidatorsOverlap {
-                total_power,
-                signed_power: voting_power,
-            }
-        );
-
-        Ok(())
-    }
-
-    fn has_sufficient_signers_overlap(
-        &self,
-        untrusted_sh: &SignedHeader,
-        untrusted_validators: &ValidatorSet,
-        calculator: &dyn VotingPowerCalculator,
-    ) -> Result<(), VerificationError> {
-        let total_power = calculator.total_power_of(untrusted_validators);
-        let signed_power = calculator
-            .voting_power_in(untrusted_sh, untrusted_validators)
-            .map_err(|e| VerificationError::ImplementationSpecific(e.to_string()))?;
-
-        ensure!(
-            signed_power * 3 > total_power * 2,
-            VerificationError::InsufficientCommitPower {
-                total_power,
-                signed_power,
-            }
-        );
-
-        Ok(())
-    }
-
-    fn valid_next_validator_set(
-        &self,
-        light_block: &LightBlock,
-        trusted_state: &TrustedState,
-    ) -> Result<(), VerificationError> {
-        ensure!(
-            light_block.signed_header.header.validators_hash
-                == trusted_state.signed_header.header.next_validators_hash,
-            VerificationError::InvalidNextValidatorSet {
-                header_next_validators_hash: light_block.signed_header.header.validators_hash,
-                next_validators_hash: trusted_state.signed_header.header.next_validators_hash,
-            }
-        );
-
-        Ok(())
-    }
-}
